added services admin and root admin flag support to include/opers.sh
changed include/opers.sh /admin definition to relfect root, services, and global admins.

changed opers.sh to generate a link block for all opers on the master server, and set the flags to o unless they are a services or global admin

added n flag support to S:lines, to support links that we don't have ssh access to.

changed example.conf to reflect this.

added root admin and services admin flags to example.conf as well

changed backup_all to be backup_sys

added a ping() command and a ping target in the main case loop

changed usage to reflect this



git-svn-id: http://gs-config.googlecode.com/svn/trunk@61 b706ddbe-c0eb-11de-acfc-b766374499fb